+++ partials/header/style/vertical-header-search.php
@@ -14,19 +14,10 @@ if ( ! defined( 'ABSPATH' ) ) {
 $post_type = get_theme_mod( 'ocean_menu_search_source', 'any' ); ?>
 
 <div id="vertical-searchform" class="header-searchform-wrap clr">
-	<form method="get" action="<?php echo esc_url( home_url( '/' ) ); ?>" class="header-searchform" aria-label="<?php esc_attr_e( 'Vertical Header Search', 'oceanwp' ); ?>">
-		<input type="search" name="s" autocomplete="off" value="" />
-		<?php
-		// If the headerSearchForm script is not disable.
-		if ( OCEAN_EXTRA_ACTIVE
-			&& class_exists( 'Ocean_Extra_Scripts_Panel' )
-			&& Ocean_Extra_Scripts_Panel::get_setting( 'oe_headerSearchForm_script' ) ) {
-			?>
-			<label><?php oceanwp_theme_strings( 'owp-string-vertical-header-search-text', 'oceanwp' ); ?></label>
-			<?php
-		}
-		?>
-		<button class="search-submit"><?php oceanwp_icon( 'search' ); ?></button>
+	<form id="verh-search" method="get" action="<?php echo esc_url( home_url( '/' ) ); ?>" class="header-searchform" aria-label="<?php esc_attr( oceanwp_theme_strings( 'owp-string-website-search-form' ) ); ?>">
+		<label for="verh-input"><?php esc_html( oceanwp_theme_strings( 'owp-string-vertical-header-search-text' ) ); ?></label>	
+		<input aria-labelledby="verh-search verh-input" id="verh-input" type="search" name="s" autocomplete="off" value="" />
+		<button class="search-submit"><?php oceanwp_icon( 'search' ); ?><span class="screen-reader-text"><?php esc_html( oceanwp_theme_strings( 'owp-string-mobile-submit-search' ) ); ?></span></button>
 		<div class="search-bg"></div>
 		<?php if ( 'any' !== $post_type ) { ?>
 			<input type="hidden" name="post_type" value="<?php echo esc_attr( $post_type ); ?>">